Added a couple languages and added tests

+++ pycodeexec/sync.py
@@ -1,6 +1,7 @@
 import docker
 from docker.models.containers import Container
 from .languages import LANGUAGES
+from threading import Thread
 
 client = docker.from_env()
 
@@ -18,17 +19,25 @@ def get_image_command(language, version):
 
 
 class Runner:
-    def __init__(self, language, version='default'):
+    def __init__(self, language, version='default', *, block=True):
         self.image, self.command = get_image_command(language, version)
-        self.container = client.containers.run(
-            self.image,
-            detach=True,
-            command="sleep 3600",
-            remove=True,
-        )  # type: Container
+        self.container = None  # type: Container
+        if block:
+            self._run_container()
+        else:
+            create_thread = Thread(target=self._run_container)
+            create_thread.start()
+
+    def _run_container(self):
+        self.container = client.containers.run(self.image, command="sleep 3600", detach=True)
+
+    @property
+    def ready(self):
+        return self.container is not None
 
     def __del__(self):
-        self.container.kill()
+        if self.container and self.container.status == 'running':
+            self.container.kill()
 
     def get_output(self, code, decode=True):
         exit_code, result = self.container.exec_run(
